+++ NEWS.md
@@ -1,3 +1,7 @@
+# PAnEn 3.7.0
+
+- Negative values in indices were removed by default. Added a parameter `debug` for backward compatibility.
+
 # PAnEn 3.6.5
 
 - Added reading forecast circular parameters in `RAnEn::readForecasts